Fix a conflict with WP-Syntax.
WP-Syntax also tries to protect certain parts of the post from WP, but runs its hooks at conflicting priorities (0 and 99, as opposed to RawHTML's 2 and 1001). The result is that the content extracted by WP-Syntax is not properly replaced back into the post. 

Fixed by detecting this situation and running our hooks betwen WP-Syntax's.

git-svn-id: https://plugins.svn.wordpress.org/raw-html/trunk@574259 b8457f37-d9ea-0310-8a92-e5e31aec5664